requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Session==0.5.0
web3==6.11.0
requests==2.31.0
Werkzeug==2.3.7
python-dotenv==1.0.0
